requirements.txt
numpy~=1.20.2
torch==1.4.0
tensorboardX~=2.1
scikit_image==0.16.2
matplotlib~=3.3.4
tqdm~=4.59.0
easydict==1.9
Pillow~=8.2.0
skimage
opencv_python
pycocotools
shapely~=1.7.1
opencv-python~=4.5.1.48
torchvision~=0.2.2.post3
packaging~=20.9
pandas~=1.2.3
seaborn~=0.11.1
yaml~=0.2.5
pyyaml~=5.4.1